@@ -2,66 +2,64 @@
   "name": "turn-based-accounting",
   "version": "0.0.0",
   "scripts": {
-    "dev": "webpack-dashboard -- webpack-dev-server",
-    "devraw": "webpack-dev-server",
-    "stats": "NODE_ENV=production webpack -p --profile --json > stats.json",
-    "dist": "npm run clean && NODE_ENV=production webpack -p",
+    "dev": "webpack-dev-server --mode development",
+    "stats": "NODE_ENV=production webpack-cli --mode production --profile --json > stats.json",
+    "dist": "yarn run clean && NODE_ENV=production webpack-cli --mode production",
     "rules": "firebase-bolt < firebase/rules/database.rules.bolt > firebase/rules/database.rules.json",
-    "deploy": "npm run test && npm run dist && firebase deploy",
+    "deploy": "yartn run test && yarn run dist && firebase deploy",
     "clean": "rimraf dist/",
     "test": "jest"
   },
   "dependencies": {
+    "@material-ui/icons": "1.0.0-beta.42",
     "@uirouter/react": "0.x.x",
-    "d3-array": "1.x.x",
-    "d3-axis": "1.x.x",
-    "d3-scale": "1.x.x",
-    "d3-selection": "1.x.x",
-    "d3-shape": "1.x.x",
+    "babel-polyfill": "6.x.x",
+    "d3-array": "1.2.1",
+    "d3-axis": "1.0.8",
+    "d3-scale": "1.0.7",
+    "d3-selection": "1.3.0",
+    "d3-shape": "1.2.0",
     "firebase": "4.x.x",
-    "firebase-bolt": "0.x.x",
-    "firebase-tools": "3.x.x",
-    "font-awesome": "4.7.0",
-    "font-awesome-webpack": "0.0.5-beta.2",
-    "material-ui": "1.0.0-beta.7",
-    "material-ui-icons": "1.0.0-alpha.3",
-    "mobile-detect": "1.3.6",
-    "prop-types": "15.x.x",
-    "react": "15.x.x",
-    "react-dom": "15.x.x",
-    "react-redux": "5.0.5",
-    "react-swipeable-views": "0.12.3",
-    "redux": "3.7.0",
-    "trianglify": "1.x.x",
+    "firebase-bolt": "0.8.2",
+    "firebase-tools": "3.18.4",
+    "material-ui": "1.0.0-beta.41",
+    "prop-types": "15.6.0",
+    "react": "16.2.0",
+    "react-dom": "16.2.0",
+    "react-firebaseui": "1.2.3",
+    "react-redux": "5.0.6",
+    "react-swipeable-views": "0.x.x",
+    "redux": "3.7.2",
+    "trianglify": "1.2.0",
     "whatwg-fetch": "2.0.3"
   },
   "devDependencies": {
     "babel-core": "6.x.x",
-    "babel-loader": "6.x.x",
-    "babel-preset-es2015": "6.x.x",
+    "babel-loader": "7.x.x",
+    "babel-preset-env": "1.x.x",
     "babel-preset-react": "6.x.x",
-    "copy-webpack-plugin": "4.0.1",
+    "copy-webpack-plugin": "4.x.x",
     "css-loader": "0.x.x",
-    "eslint": "3.19.0",
-    "eslint-config-airbnb": "15.0.1",
-    "eslint-loader": "1.x.x",
-    "eslint-plugin-import": "2.3.0",
-    "eslint-plugin-jsx-a11y": "5.0.3",
-    "eslint-plugin-react": "7.1.0",
-    "extract-text-webpack-plugin": "2.1.0",
-    "file-loader": "0.11.2",
-    "html-webpack-plugin": "2.x.x",
-    "jest": "20.0.4",
-    "less": "2.7.2",
+    "eslint": "4.19.1",
+    "eslint-config-airbnb": "16.1.0",
+    "eslint-loader": "2.x.x",
+    "eslint-plugin-import": "2.11.0",
+    "eslint-plugin-jsx-a11y": "6.0.3",
+    "eslint-plugin-react": "7.7.0",
+    "extract-text-webpack-plugin": "3.x.x",
+    "file-loader": "1.x.x",
+    "html-webpack-plugin": "3.x.x",
+    "jest": "22.x.x",
+    "less": "2.x.x",
     "node-sass": "4.x.x",
     "rimraf": "2.x.x",
-    "sass-loader": "6.x.x",
+    "sass-loader": "7.x.x",
     "style-loader": "0.x.x",
-    "url-loader": "0.5.9",
-    "webpack": "2.x.x",
-    "webpack-dashboard": "0.x.x",
-    "webpack-dev-server": "2.x.x",
-    "webpack-merge": "2.x.x",
+    "url-loader": "0.x.x",
+    "webpack": "4.x.x",
+    "webpack-cli": "2.x.x",
+    "webpack-dev-server": "3.x.x",
+    "webpack-merge": "4.x.x",
     "webpack-notifier": "1.x.x"
   }
 }
