@@ -3,20 +3,25 @@ import Grid from 'material-ui/Grid';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
-const Amount = require('../components/amount/index.jsx');
-const DateScroller = require('./components/date-scroller.jsx');
-const RecurrentList = require('../components/recurrent/list.jsx');
-const PunctualList = require('../components/punctual/list.jsx');
+import Amount from '../components/amount/index';
+import DateScroller from './components/date-scroller';
+import RecurrentList from '../components/recurrent/list';
+import PunctualList from '../components/punctual/list';
 
-const DB = require('../firebase/database.js');
-const DateUtils = require('../utils/date.js');
-const Actions = require('../actions/index.js');
+import { getPunctualRef, getRecurrentRef } from '../firebase/database';
+import { formatDate } from '../utils/date';
+import { updatePunctual, updateRecurrent } from '../actions/index';
 
 const styles = {
   heightHeader: {
     height: '150px',
     marginTop: '20px',
     marginBottom: '20px',
+  },
+  root: {
+    width: '100%',
+    marginTop: '10px',
+    paddingLeft: '10px',
   },
 };
 
@@ -37,17 +42,17 @@ class Bilan extends React.Component {
   }
 
   componentDidMount() {
-    this.enablePunctualEntries(this.props.resolves.date, this.props.type);
-    this.enableRecurrentEntries(this.props.resolves.date, this.props.type);
+    this.enablePunctualEntries(this.props.date, this.props.type);
+    this.enableRecurrentEntries(this.props.date, this.props.type);
   }
 
   componentWillReceiveProps(nextProps) {
-    if (this.props.resolves.date !== nextProps.resolves.date ||
+    if (this.props.date !== nextProps.date ||
       this.props.type !== nextProps.type) {
-      this.disablePunctualEntries(this.props.resolves.date, this.props.type);
-      this.enablePunctualEntries(nextProps.resolves.date, nextProps.type);
+      this.disablePunctualEntries(this.props.date, this.props.type);
+      this.enablePunctualEntries(nextProps.date, nextProps.type);
 
-      this.filterRecurrentEntries(nextProps.resolves.date, nextProps.type);
+      this.filterRecurrentEntries(nextProps.date, nextProps.type);
     }
   }
 
@@ -56,25 +61,23 @@ class Bilan extends React.Component {
   }
 
   componentWillUnmount() {
-    this.disablePunctualEntries(this.props.resolves.date, this.props.type);
+    this.disablePunctualEntries(this.props.date, this.props.type);
     this.disableRecurrentEntries();
   }
 
   enablePunctualEntries(date, type) {
-    this.callbackPunctual = DB.getPunctualRef(date, type).on('value', snapshot =>
-      this.context.store.dispatch(Actions.updatePunctual(snapshot.val() || {}))
-
-    );
+    this.callbackPunctual = getPunctualRef(date, type).on('value', snapshot =>
+      this.context.store.dispatch(updatePunctual(snapshot.val() || {})));
   }
 
   disablePunctualEntries(date, type) {
     if (this.callbackPunctual) {
-      DB.getPunctualRef(date, type).off('value', this.callbackPunctual);
+      getPunctualRef(date, type).off('value', this.callbackPunctual);
     }
   }
 
   enableRecurrentEntries(date, type) {
-    this.callbackRecurrent = DB.getRecurrentRef().on('value', (snapshot) => {
+    this.callbackRecurrent = getRecurrentRef().on('value', (snapshot) => {
       this.recurrentEntries = snapshot.val() || {};
       this.filterRecurrentEntries(date, type);
     });
@@ -82,13 +85,13 @@ class Bilan extends React.Component {
 
   disableRecurrentEntries() {
     if (this.callbackRecurrent) {
-      DB.getRecurrentRef().off('value', this.callbackRecurrent);
+      getRecurrentRef().off('value', this.callbackRecurrent);
     }
   }
 
   filterRecurrentEntries(date, type) {
     const res = {};
-    const { startDate, endDate } = DateUtils.formatDate(date, type);
+    const { startDate, endDate } = formatDate(date, type);
 
     Object.keys(this.recurrentEntries).forEach((i) => {
       if (this.recurrentEntries[i].startDate && this.recurrentEntries[i].startDate > endDate) {
@@ -103,12 +106,19 @@ class Bilan extends React.Component {
     });
 
 
-    this.context.store.dispatch(Actions.updateRecurrent(res));
+    this.context.store.dispatch(updateRecurrent(res));
   }
 
   render() {
     return (
-      <Grid container direction="row" justify="space-between" align="flex-start">
+      <Grid
+        container
+        direction="row"
+        justify="space-between"
+        align="flex-start"
+        spacing={16}
+        style={styles.root}
+      >
         <Grid item md={6} xs={12} style={styles.heightHeader}>
           <Amount />
         </Grid>
@@ -127,9 +137,7 @@ class Bilan extends React.Component {
 }
 
 Bilan.propTypes = {
-  resolves: PropTypes.shape({
-    date: PropTypes.instanceOf(Date).isRequired,
-  }).isRequired,
+  date: PropTypes.instanceOf(Date).isRequired,
   type: PropTypes.oneOf(['day', 'week', 'month', 'year']).isRequired,
 };
 
@@ -137,4 +145,4 @@ Bilan.contextTypes = {
   store: PropTypes.object.isRequired,
 };
 
-module.exports = connect(mapStateToProps)(Bilan);
+export default connect(mapStateToProps)(Bilan);
