@@ -1,31 +1,33 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { UIRouter, UIView, pushStateLocationPlugin } from '@uirouter/react';
+import { UIRouter, UIView, hashLocationPlugin } from '@uirouter/react';
+import CssBaseline from 'material-ui/CssBaseline';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
 
-const theme = require('./material-ui-theme.js');
+import theme from './material-ui-theme';
 
-const Router = require('./routes.js');
+import reducers from './reducers/index';
+import { startLocalization } from './utils/l10n';
+import * as Router from './routes';
 
-require('font-awesome-webpack');
-
-require('./index.scss');
-
-const store = createStore(require('./reducers/index.js'));
-require('./utils/l10n.js').start(store);
+// require('./index.scss');
+const store = createStore(reducers);
+startLocalization(store);
 
 ReactDOM.render(
   <MuiThemeProvider theme={theme}>
+    <CssBaseline />
     <Provider store={store}>
       <UIRouter
-        plugins={[pushStateLocationPlugin]}
+        plugins={[hashLocationPlugin]}
         states={Router.getStates()}
         config={Router.getConfig}
       >
         <UIView />
       </UIRouter>
     </Provider>
-  </MuiThemeProvider>
-, document.getElementById('tba-app'));
+  </MuiThemeProvider>,
+  document.getElementById('tba-app'),
+);
