@@ -3,7 +3,7 @@ import teal from 'material-ui/colors/teal';
 import orange from 'material-ui/colors/orange';
 import red from 'material-ui/colors/red';
 
-module.exports = createMuiTheme({
+export default createMuiTheme({
   palette: {
     primary: teal,
     secondary: orange,
