@@ -4,7 +4,7 @@ const UPDATE_RECURRENT = 'UPDATE_RECURRENT';
 const UPDATE_DATE_TYPE = 'UPDATE_DATE_TYPE';
 const UPDATE_CURRENCY = 'UPDATE_CURRENCY';
 
-module.exports = {
+export {
   UPDATE_LOCALE,
   UPDATE_PUNCTUAL,
   UPDATE_RECURRENT,
