@@ -1,17 +1,22 @@
+'use strict'
+
 const path = require('path');
 
 const webpack = require('webpack');
 const webpackMerge = require('webpack-merge');
 const htmlWebpackPlugin = require('html-webpack-plugin');
-const DashboardPlugin = require('webpack-dashboard/plugin');
 const webpackNotifierPlugin = require('webpack-notifier');
 const webpackExtractTextPlugin = require('extract-text-webpack-plugin');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
 
 const SRC_DIR = path.resolve(__dirname, 'src');
 
+function srcPath(p) {
+  return path.resolve(SRC_DIR, p);
+}
+
 const htmlWebpackPluginConfig = {
-  template: path.resolve(SRC_DIR, 'index.html'),
+  template: srcPath('index.html'),
   filename: 'index.html',
   minify: {
     removeComments: true,
@@ -19,16 +24,16 @@ const htmlWebpackPluginConfig = {
   },
 };
 
-function newWebpackTextPlugin(filename) {
-  return new webpackExtractTextPlugin({ filename, allChunks: true });
-}
+// function newWebpackTextPlugin(filename) {
+//   return new webpackExtractTextPlugin({ filename, allChunks: true });
+// }
 
-let bundleCSS;
-if (process.env.NODE === 'production') {
-  bundleCSS = newWebpackTextPlugin('main.[contenthash].css');
-} else {
-  bundleCSS = newWebpackTextPlugin('main.css');
-}
+// let bundleCSS;
+// if (process.env.NODE === 'production') {
+//   bundleCSS = newWebpackTextPlugin('main.[contenthash].css');
+// } else {
+//   bundleCSS = newWebpackTextPlugin('main.css');
+// }
 
 //
 // Common configuration
@@ -36,11 +41,11 @@ if (process.env.NODE === 'production') {
 const common = {
   output: {
     path: path.resolve(__dirname, 'dist'),
-    // publicPath: '/tba/',
+    publicPath: '/',
   },
   entry: {
-    vendor: path.resolve(SRC_DIR, 'vendor.js'),
-    main: path.resolve(SRC_DIR, 'index.jsx'),
+    // vendor: srcPath('vendor.js'),
+    main: srcPath('index.jsx'),
   },
   module: {
     rules: [
@@ -48,31 +53,26 @@ const common = {
         { loader: 'babel-loader' },
         { loader: 'eslint-loader' },
       ] },
-      { test: /\.scss$/, include: SRC_DIR, loader:
-        bundleCSS.extract({ use: [
-          { loader: 'css-loader' },
-          { loader: 'sass-loader' }
-        ] })
-      },
-      // font-awesome
-      // TODO: extract it to the vendor css
-      { test: /\.woff(2)?(\?v=[a-z0-9]\.[a-z0-9]\.[a-z0-9])?$/, include: /font-awesome/, use: [
-        { loader: 'url-loader?limit=100000&mimetype=application/font-woff' }
-      ] },
-      { test: /\.(ttf|eot|svg)(\?v=[a-z0-9]\.[a-z0-9]\.[a-z0-9])?$/, include: /font-awesome/, use: [
-        { loader: 'file-loader' }
-      ] },
+      // { test: /\.scss$/, include: SRC_DIR, loader:
+      //   bundleCSS.extract({ use: [
+      //     { loader: 'css-loader' },
+      //     { loader: 'sass-loader' }
+      //   ] })
+      // },
     ],
   },
+  resolve: {
+    extensions: ['.js', '.jsx'],
+  },
   plugins: [
-    bundleCSS,
+    // bundleCSS,
     new htmlWebpackPlugin(htmlWebpackPluginConfig),
     new CopyWebpackPlugin([
       { from: 'src/lang', to: 'lang', transform: (content => JSON.stringify(JSON.parse(content))) },
     ]),
-    new webpack.optimize.CommonsChunkPlugin({
-      names: ['vendor', 'manifest'],
-    }),
+    // new webpack.optimize.CommonsChunkPlugin({
+    //   names: ['vendor', 'manifest'],
+    // }),
   ],
 };
 
@@ -81,9 +81,12 @@ if (process.env.NODE_ENV === 'production') {
   // Production configuration
   //
   module.exports = webpackMerge(common, {
-    devtool: 'cheap-module-source-map',
+    devtool: 'source-map',
     output: {
       filename: '[name].[chunkhash].js',
+    },
+    entry: {
+      main: ['babel-polyfill', srcPath('index.jsx')],
     },
   });
 } else {
@@ -91,21 +94,17 @@ if (process.env.NODE_ENV === 'production') {
   // Development / Test configuration
   //
   module.exports = webpackMerge(common, {
-    devtool: 'cheap-eval-source-map',
+    devtool: 'cheap-module-source-map',
     output: {
       filename: '[name].js',
     },
     devServer: {
       quiet: false,
       noInfo: false,
-      historyApiFallback: true,
+      historyApiFallback: false,
       port: 3000,
     },
-    externals: {
-      'Config': JSON.stringify(require('./config/config.json')),
-    },
     plugins: [
-      new DashboardPlugin(),
       new webpackNotifierPlugin({ alwaysNotify: true }),
     ]
   });
